#21: move currency card background to frame
